borges-tool: fix bugs in delete and queue
* connect to amqp queue in CLI
* delete should not remove references is running with --dry

+++ tool/siva.go
@@ -136,7 +136,11 @@ func (s *Siva) deleteDatabase(ctx context.Context, init string) error {
 		}
 	}
 
-	return s.db.DeleteReferences(ctx, init)
+	if !s.dry {
+		return s.db.DeleteReferences(ctx, init)
+	}
+
+	return nil
 }
 
 func (s *Siva) deleteFilesystem(ctx context.Context, init string) error {
